:sparkles: feat: 로딩스피너 추가, 댓글내용에 변동이 없으면 post요청이 실행되지 않게 수정
