+++ FE/.env
@@ -1,3 +1,3 @@
 NEXT_PUBLIC_API_MOCKING=enabled
-NEXT_PUBLIC_SERVER_URL=''
+NEXT_PUBLIC_SERVER_URL='https://a683-222-233-138-154.jp.ngrok.io'
 NEXT_PUBLIC_DEFAULT_IMG='https://scontent-gmp1-1.xx.fbcdn.net/v/t1.18169-9/527016_499021583525593_732357164_n.jpg?_nc_cat=107&ccb=1-7&_nc_sid=09cbfe&_nc_ohc=CdbnqyyFWXkAX_obHCp&_nc_ht=scontent-gmp1-1.xx&oh=00_AfDBQSsLnCXMKoAPnGFrOeSBgvMp__vgjXLEqmtS6etfcw&oe=63F1DB6A'
