+++ FE/src/components/VoteList/sidebar.tsx
