+++ FE/src/utils/userToken.ts
@@ -4,7 +4,6 @@ import SessionStorage from '../constants/sessionstorage';
 export const getToken = () => {
   const localToken = LocalStorage.getItem('accesstoken');
   const sessionToken = SessionStorage.getItem('accesstoken');
-  // console.log(localToken, sessionToken);
 
   if (localToken !== undefined && localToken !== null) {
     return localToken;
@@ -13,10 +12,4 @@ export const getToken = () => {
   } else {
     return;
   }
-};
-
-export const deleteToken = () => {
-  LocalStorage.removeItem('accesstoken');
-  LocalStorage.removeItem('refreshtoken');
-  SessionStorage.removeItem('accesstoken');
 };
